+++ javascope/jScope/WaveformMetrics.java
@@ -99,14 +99,12 @@ public class WaveformMetrics
         xrange = xmax - xmin;
         yrange = ymax - ymin;
 
-        if (xrange <= 0)
-        {
-            xrange = (double) 1E-10;
+        if (xrange <= 0) {
+            xrange = 1E-10;
             x_offset = 0.5;
         }
-        if (yrange <= 0)
-        {
-            yrange = (double) 1E-10;
+        if (yrange <= 0) {
+            yrange = 1E-10;
         }
     }
 
@@ -216,8 +214,7 @@ public class WaveformMetrics
 
     final public double XValue(int x, Dimension d)
     {
-        double ris = (double) ( ( (x - 0.5) / d.width - x_offset) * xrange /
-                               x_range + xmin);
+        double ris = ((x - 0.5) / d.width - x_offset) * xrange / x_range + xmin;
         if (x_log)
             return Math.exp(LOG10 * ris);
         else
@@ -226,8 +223,7 @@ public class WaveformMetrics
 
     final public double YValue(int y, Dimension d)
     {
-        double ris = (double) (ymax -
-                               ( (y - 0.5) / d.height) * yrange / y_range);
+        double ris = ymax - ((y - 0.5) / d.height) * yrange / y_range;
         if (y_log)
             return Math.exp(LOG10 * ris);
         else
